Initializes clinspacy. This function is optional to run but gives you more control over the parameters used by scispacy at initiation. If you do not run this function, it will be run with default parameters the first time that any of the package functions are run.

Initializes clinspacy. This function is optional to run but gives you more control over the parameters used by scispacy at initiation. If you do not run this function, it will be run with default parameters the first time that any of the package functions are run.

clinspacy_init(
  miniconda = TRUE,
  use_linker = FALSE,
  linker_threshold = 0.99,
  ...
)

Arguments

miniconda
Defaults to TRUE, which results in miniconda being installed (~400 MB) and configured with the "clinspacy" conda environment. If you want to override this behavior, set miniconda to FALSE and specify an alternative environment using use_python() or use_conda().

use_linker
Defaults to FALSE. To turn on the UMLS linker, set this to TRUE.

linker_threshold
Defaults to 0.99. This arguemtn is only relevant if use_linker is set to TRUE. It refers to the confidence threshold value used by the scispacy UMLS entity linker. Note: This can be lower than the threshold from clinspacy_init). The linker_threshold can only be set once per session.

...
Additional settings available from: https://github.com/allenai/scispacy.

Value

No return value.