Cui2vec concept embeddings

This dataset contains Unified Medical Langauge System (UMLS) concept embeddings from Andrew Beam's cui2vec R package. There are 500 embeddings included for each concept.

cui2vec_embeddings

Format

A data frame with 109053 rows and 501 variables:

cui

A Unified Medical Language System (UMLS) Concept Unique Identifier (CUI)

emb_001

Concept embedding vector #1

emb_002

Concept embedding vector #2

...

and so on...

emb_500

Concept embedding vector #500

Source

https://figshare.com/s/00d69861786cd0156d81

Details

This dataset is not viewable until it has been downloaded, which will occur the very first time you run clinspacy_init() after installing this package.

Citation

Beam, A.L., Kompa, B., Schmaltz, A., Fried, I., Griffin, W, Palmer, N.P., Shi, X., Cai, T., and Kohane, I.S.,, 2019. Clinical Concept Embeddings Learned from Massive Sources of Multimodal Medical Data. arXiv preprint arXiv:1804.01486.

License

This data is made available under a CC BY 4.0 license. The only change made to the original dataset is the renaming of columns.