This function binds columns containing concept embeddings for concepts with which scispacy has 99 percent confidence of being present with values containing frequencies. Negated concepts, as identified by negspacy's NegEx implementation, are ignored and do not count towards the embeddings. The concept embeddings are derived from the cui2vec_embeddings dataset included with this package.

The embeddings are derived from Andrew Beam's cui2vec R package.

bind_clinspacy_embeddings(
  df,
  text,
  type = "scispacy",
  num_embeddings = 200,
  ...
)

Arguments

df
A data frame.

text
A character string containing the name of the column to process.

type
The type of embeddings to return. One of cui2vec and scispacy. Whereas cui2vec embeddings require the UMLS linker to be enabled, the scispacy embeddings do not. Defaults to scispacy.

num_embeddings
The number of embeddings to return. This must be a number 1 through 200 for scispacy embeddings and 1 through 500 for cui2vec embeddings. Defaults to 200.

...
Arguments passed down to clinspacy

Value

A data frame containing the original data frame as well as additional column names for each UMLS concept unique identifer found with values containing frequencies.

Details

Citation

Beam, A.L., Kompa, B., Schmaltz, A., Fried, I., Griffin, W, Palmer, N.P., Shi, X., Cai, T., and Kohane, I.S.,, 2019. Clinical Concept Embeddings Learned from Massive Sources of Multimodal Medical Data. arXiv preprint arXiv:1804.01486.

License

This data is made available under a CC BY 4.0 license. The only change made to the original dataset is the renaming of columns.

Examples

data(mtsamples)
mtsamples_with_cuis = bind_clinspacy(mtsamples[1:5,], text = 'description')

#> Processing... A 23-year-old white female presents with complaint of allergies.

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======================                                               |  33%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================================| 100%

#> Processing... Consult for laparoscopic gastric bypass.

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

#> Processing... Consult for laparoscopic gastric bypass.

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

#> Processing... 2-D M-Mode. Doppler.

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

#> Processing... 2-D Echocardiogram

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

str(mtsamples_with_cuis)

#> 'data.frame':	5 obs. of  15 variables:
#>  $ note_id                    : int  1 2 3 4 5
#>  $ description                : chr  "A 23-year-old white female presents with complaint of allergies." "Consult for laparoscopic gastric bypass." "Consult for laparoscopic gastric bypass." "2-D M-Mode. Doppler." ...
#>  $ medical_specialty          : chr  "Allergy / Immunology" "Bariatrics" "Bariatrics" "Cardiovascular / Pulmonary" ...
#>  $ sample_name                : chr  "Allergic Rhinitis" "Laparoscopic Gastric Bypass Consult - 2" "Laparoscopic Gastric Bypass Consult - 1" "2-D Echocardiogram - 1" ...
#>  $ transcription              : chr  "SUBJECTIVE:,  This 23-year-old white female presents with complaint of allergies.  She used to have allergies w"| __truncated__ "PAST MEDICAL HISTORY:, He has difficulty climbing stairs, difficulty with airline seats, tying shoes, used to p"| __truncated__ "HISTORY OF PRESENT ILLNESS: , I have seen ABC today.  He is a very pleasant gentleman who is 42 years old, 344 "| __truncated__ "2-D M-MODE: , ,1.  Left atrial enlargement with left atrial diameter of 4.7 cm.,2.  Normal size right and left "| __truncated__ ...
#>  $ keywords                   : chr  "allergy / immunology, allergic rhinitis, allergies, asthma, nasal sprays, rhinitis, nasal, erythematous, allegr"| __truncated__ "bariatrics, laparoscopic gastric bypass, weight loss programs, gastric bypass, atkin's diet, weight watcher's, "| __truncated__ "bariatrics, laparoscopic gastric bypass, heart attacks, body weight, pulmonary embolism, potential complication"| __truncated__ "cardiovascular / pulmonary, 2-d m-mode, doppler, aortic valve, atrial enlargement, diastolic function, ejection"| __truncated__ ...
#>  $ 2-D                        : num  0 0 0 0 1
#>  $ 2-D M-Mode                 : num  0 0 0 1 0
#>  $ Consult                    : num  0 1 1 0 0
#>  $ Doppler                    : num  0 0 0 1 0
#>  $ Echocardiogram             : num  0 0 0 0 1
#>  $ allergies                  : num  1 0 0 0 0
#>  $ complaint                  : num  1 0 0 0 0
#>  $ laparoscopic gastric bypass: num  0 1 1 0 0
#>  $ white female               : num  1 0 0 0 0